send mint tx as a context trait method
